cleanup, examples and docs updated
