+++ examples/audio_capture.livemd
@@ -6,31 +6,17 @@ Logger.configure(level: :error)
 
 Mix.install([
   {:membrane_core, "~> 1.0"},
-  {:membrane_file_plugin, "~> 0.17.0"},
   {:membrane_kino_plugin, path: ".."},
-  {:membrane_opus_plugin, "~> 0.20.2"},
-  {:membrane_aac_fdk_plugin, "~> 0.18.8"},
-  {:membrane_aac_plugin, "~> 0.18.1"},
-  {:membrane_wav_plugin, "~> 0.10.1"},
-  {:membrane_generator_plugin, "~> 0.10.0"},
-  {:membrane_raw_audio_format, "~> 0.12.0"},
   {:membrane_matroska_plugin, "~> 0.5.1"},
-  {:membrane_portaudio_plugin, "~> 0.19.2"},
-  {:membrane_ffmpeg_swresample_plugin, "~> 0.20.2"},
-  {:membrane_h264_ffmpeg_plugin, "~> 0.31.6"},
-  {:membrane_fake_plugin, "~> 0.11.0"},
-  {:membrane_raw_video_format, "~> 0.3.0"}
-
-  # {:membrane_ffmpeg_swscale_plugin, "~> 0.15.1"},
-  # {:membrane_funnel_plugin, "~> 0.9.0"},
-  # {:membrane_matroska_plugin, "~> 0.5.1"},
+  {:membrane_opus_plugin, "~> 0.20.2"},
+  {:membrane_aac_fdk_plugin, "~> 0.18.8"}
 ])
 ```
 
 ## Pipeline definition
 
 ```elixir
-alias Membrane.RCPipeline, as: RC
+alias Membrane.RCPipeline
 import Membrane.ChildrenSpec
 
 alias Membrane.{
@@ -54,7 +40,8 @@ kino_player =
   )
 
 play_kino_audio =
-  child(:input, %Membrane.Kino.Input.AudioSource{kino: kino_input})
+  child(:input, %Membrane.Kino.Input.Bin.Source{kino: kino_input})
+  |> via_out(:audio)
   |> child(:from_opus, Opus.Decoder)
   |> child(:to_aac, AAC.FDK.Encoder)
   |> via_in(:audio)
@@ -66,8 +53,8 @@ play_kino_audio =
 ## Playing pipeline
 
 ```elixir
-pipeline = RC.start!()
-RC.exec_actions(pipeline, spec: play_kino_audio)
+pipeline = RCPipeline.start!()
+RCPipeline.exec_actions(pipeline, spec: play_kino_audio)
 
 Kino.render(kino_input)
 kino_player
