+++ lib/assets/input/main.js
@@ -102,8 +102,8 @@ export async function init(ctx, info) {
       };
 
       await video_track.applyConstraints(video_stream_constraints);
+      
       let video_settings = video_track.getSettings();
-      console.log(video_settings)
       ctx.pushEvent('framerate', video_settings.frameRate);
 
       const options = {
