+++ lib/membrane/kino/input/input_bin_source.ex
@@ -1,6 +1,6 @@
-defmodule Membrane.Kino.Input.SourceBin do
+defmodule Membrane.Kino.Input.Bin.Source do
   @moduledoc """
-  This module provides a video input source compatible with the Livebook environment.
+  This module provides audio and/or video input source compatible with the Livebook environment.
 
   Livebook handles multimedia and specific media by using the Kino library and its extensions.
   This module integrate special `Membrane.Kino.Input` element into the Membrane pipeline.
@@ -31,7 +31,7 @@ defmodule Membrane.Kino.Input.SourceBin do
   @impl true
   def handle_init(_ctx, options) do
     spec =
-      child(:remote_stream, %Kino.Input.Source.RemoteStream{kino: options.kino})
+      child(:remote_stream, %Kino.Input.Source{kino: options.kino})
 
     {[spec: spec], %{framerate: nil, tracks: %{}}}
   end
@@ -45,7 +45,8 @@ defmodule Membrane.Kino.Input.SourceBin do
         generate_best_effort_timestamps: %{framerate: {framerate, 1}}
       })
       |> get_child(:funnel_video)
-    {[spec: spec],  %{state | framerate: framerate}}
+
+    {[spec: spec], %{state | framerate: framerate}}
   end
 
   @impl true
@@ -65,27 +66,25 @@ defmodule Membrane.Kino.Input.SourceBin do
   end
 
   @impl true
-  def handle_pad_added({_, name, _ref} = pad, _ctx, state) do
-    # todo check if pads already exist
-    spec = case name do
-      :audio ->
-        [
-          get_child(:remote_stream)
-          |> via_out(:audio)
-          |> child(:demuxer, Matroska.Demuxer),
-          child(:funnel_audio, Funnel) |> bin_output(pad),
-        ]
+  def handle_pad_added({_membrane_pad, name, _ref} = pad, _ctx, state) do
+    spec =
+      case name do
+        :audio ->
+          [
+            get_child(:remote_stream)
+            |> via_out(:audio)
+            |> child(:demuxer, Matroska.Demuxer),
+            child(:funnel_audio, Funnel) |> bin_output(pad)
+          ]
 
-      :video ->
-        child(:funnel_video, Funnel)
-        |> bin_output(pad)
+        :video ->
+          child(:funnel_video, Funnel)
+          |> bin_output(pad)
 
-      true ->
-        raise "adding unknown pad"
-    end
+        true ->
+          raise "adding unknown pad"
+      end
 
     {[spec: spec], %{state | tracks: Map.put(state.tracks, pad, Track)}}
-
   end
-
 end
